Added toggle for ignoring undefined tech level projects when doing cost calcuations
This is somewhat redundant with the tech tab one since in the base game the only undefined tech level projects are Anomaly. However, this will work on any mods that for whatever reason don't set a tech level on their projects and also is a secondary and more simple way to handle the Anomaly tab.